requirements.txt
PyYAML
icecream
typer
rich